@@ -11,14 +11,14 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.1",
-    "@angular/common": "^6.1.1",
-    "@angular/compiler": "^6.1.1",
-    "@angular/core": "^6.1.1",
-    "@angular/forms": "^6.1.1",
-    "@angular/platform-browser": "^6.1.1",
-    "@angular/platform-browser-dynamic": "^6.1.1",
-    "@angular/router": "^6.1.1",
+    "@angular/animations": "^8.1.2",
+    "@angular/common": "^8.1.2",
+    "@angular/compiler": "^8.1.2",
+    "@angular/core": "^8.1.2",
+    "@angular/forms": "^8.1.2",
+    "@angular/platform-browser": "^8.1.2",
+    "@angular/platform-browser-dynamic": "^8.1.2",
+    "@angular/router": "^8.1.2",
     "core-js": "^2.5.4",
     "rxjs": "^6.0.0",
     "zone.js": "~0.9.1"
@@ -27,8 +27,8 @@
     "@angular-devkit/build-angular": "~0.801.2",
     "@angular-devkit/build-ng-packagr": "~0.801.2",
     "@angular/cli": "~8.1.2",
-    "@angular/compiler-cli": "^6.1.1",
-    "@angular/language-service": "^6.1.1",
+    "@angular/compiler-cli": "^8.1.2",
+    "@angular/language-service": "^8.1.2",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
@@ -46,6 +46,6 @@
     "tsickle": "^0.35.0",
     "tslib": "^1.9.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.7.2"
+    "typescript": "~3.4.5"
   }
 }
